backend/requirements.txt
fastapi
uvicorn
cohere
qdrant-client
sqlalchemy
psycopg[binary]
pydantic
python-dotenv
httpx
openai-agents
markdown
beautifulsoup4
psycopg